+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 * 【ai     】      修复`AIConfigBuilder`中方法名拼写错误（pr#1382@Gitee）
 * 【core   】      修复`StrBuilder`charAt越界判断错误（pr#4094@Github）
 * 【dfa    】      修复`WordTree.addWord`末尾为特殊字符导致的无法匹配问题（pr#4092@Github）
-* 【core   】      修复`ServiceLoaderUtil.loadFirstAvailable`在JDK24+后未捕获异常导致的报错问题（issue#ID0952@Gitee）
+* 【core   】      修复`ServiceLoaderUtil.loadFirstAvailable`在JDK24+后未捕获异常导致的报错问题（pr#4098@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.40(2025-08-26)